refactor: Convert latlong Direwolf tests to direct Go test functions
